+++ src/extension.ts
@@ -3,6 +3,9 @@ import { Emulator } from "./emulator/Emulator";
 
 import { contextMenu } from "./Context_Menu/contextMenu";
 import { xsd } from "./xsd/xsd";
+import { snippet } from "./snippets/snippets";
+import { xml } from "./xml/extension";
+
 
 export function activate(context: vscode.ExtensionContext) {
   console.log(
@@ -12,8 +15,8 @@ export function activate(context: vscode.ExtensionContext) {
   new Emulator(context);
   new contextMenu(context);
   new xsd(context);
-
- 
+  new snippet(context);
+  xml(context)
 }
 
-export function deactivate() {}
+export function deactivate() { }